demo/fpga-risc-cpu/requirements.txt
cocotb==1.7.1
cocotbext-uart==0.1.2
pytest==7.1.3
pyserial==3.5
python-constraint
https://github.com/chipsalliance/f4pga/archive/e1cd038f06c7161b27afd0073fb507da2b8e5a9e.zip#subdirectory=f4pga
intervaltree
junit-xml
numpy
openpyxl
ordered-set
parse
progressbar2
pyjson5
pytest
pyyaml
scipy>=1.2.1
simplejson
sympy
textx
yapf==0.24.0
fasm
git+https://github.com/f4pga/prjxray.git@ae546d6b7648bf4df9cf63f0b25b2028b5623c43#egg=prjxray
git+https://github.com/chipsalliance/f4pga-xc-fasm.git@25dc605c9c0896204f0c3425b52a332034cf5e5c#egg=xc-fasm
z3-solver